pyproject.toml
[project]
name = "ahrefs-geo-bot"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aiogram>=3.18.0",
    "curl-cffi>=0.10.0",
    "loguru>=0.7.3",
    "openpyxl>=3.1.5",
    "pydantic-settings>=2.8.1",
    "python-dotenv>=1.0.1",
]
